requirements.txt
faker
pyspark
dash
plotly
psycopg2-binary
kafka-python
apscheduler
python-dotenv
